[core] Move common code from FileSystemCatalog and HiveCatalog into AbstractCatalog
